files/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
openai==1.35.3
google-auth==2.29.0
google-api-python-client==2.134.0
dateparser==1.2.0
python-multipart==0.0.9
